Made entering buildings work by touching door or player when player close
also put all code to check if player is in front of door, and to move
player to url of the said door into a helper function called
attemptEnteringDoor.

When the player is tapped, all doors are checked to see if the player is
close to one, if he is then he will enter it. When a door is tapped, it
is checked whether or not the player is close to that door, if he is
then he will enter it.